Add logic path to disable Managed Identity
Managed Identity requires a "Owner" service principal
to create and manage role assignments for the managed
identities (VMs). This presents the problem of creating
a service principal with such permissions (Application credential).

Creating such application credentials present a grave security
risk which is mitigated with this change. This change
allows Charmed Kubernetes to use the service principal directly,
as opposed to a service principal which assigns roles to VMs
the VMs utilise a SP that already has the roles required.